==> 1_Preproceso.ipynb <==
# 1. Preproceso

## 1.1 Imports y carga de dataset

Al comienzo de este notebook, se ejecutará todos los imports usados en él, y el dataset utilizado para llevar a cabo el modelo.

#Imports y carga del dataset

from nltk.corpus import stopwords
from nltk.tokenize import word_tokenize
from nltk.stem import WordNetLemmatizer
import pandas as pd
import nltk
import math
import ast
import io
import re
import warnings
warnings.filterwarnings('ignore')
import numpy as np

nltk.download("stopwords")
nltk.download("punkt")
nltk.download("wordnet")

#pd.set_option("display.max_rows", None)

with open("movies_dataset/movies_metadata.csv", "r", encoding='utf-8') as f:
    df_raw = pd.read_csv(f, encoding="utf-8")
    print("Dataset loaded")

## 1.2 Eliminación de columnas no usadas

Vamos a procesar todas las columnas del dataset y construir un dataframe nuevo.

Las columnas que hemos considerado innecesarias para incluir son:

* ``homepage``
* ``imdb_id``
* ``original_title``
* ``poster_path``
* ``status``
* ``video``
* ``id``
* ``popularity``

The column ``id`` won't be removed, in order to know and to identify what movie belongs to each row in the preprocessed data.

exclude = ["homepage", "imdb_id", "original_title", "poster_path", "status", "video", "popularity"] 
df = df_raw.loc[:, df_raw.columns.difference(exclude)]
print("Columnas excluidas")

Una vez eliminada estas columnas, se comprueba que columnas son las que se van a procesar y cuales tienen algún valor nulo.

#Comprobación de valores nulos

print(pd.isnull(df).any())
print()
print("Filas totales:", len(df))
print("Filas con algún valor nulo:", len([x for x in df.isnull().any(axis=1) if x]))

## 1.3 Procesamiento de columnas

### 1.3.1 Adult

Aqui observamos que hay valores no booleanos que debemos eliminar, y además observamos que las películas con valor True son muy escasas en comparación con las que tienen valor False, por lo que se procederá a eliminar las filas anómalas y quitar la columna ``adult``.

print(df["adult"].value_counts())
print()

drop_list = []
for i in range(len(df)):
    if df["adult"][i] != "False":
        drop_list.append(i)

df.drop(index=drop_list, inplace=True)

print(df["adult"].value_counts())

df.drop(["adult"], axis=1, inplace=True)

#No añadimos la columna al nuevo dataframe

### 1.3.2 Title

Antes de procesar esta columna, se pasará a definir una serie de funciones que nos servirán para procesar aquellas carácterísticas que están formadas por lenguaje natural.

# Definicion de funciones para el procesamiento de lenguaje natural

replaces = [("i'm", "i am"), ("you're", "you are"), ("we're", "we are"), ("they're", "they are"),
            ("that's", "that is"), ("there's", "there is"), ("who's", "who is"), ("what's", "what is"),
            
            ("isn't", "is not"), ("aren't", "are not"), ("can't", "cannot"), ("haven't", "have not"),
            ("won't", "will not"), ("doesn't", "does not"), ("don't", "do not"), ("didn't", "did not"),
            ("couldn't", "could not"), ("shouldn't", "should not"), ("wasn't", "was not"), ("weren't", "were not"),
            ("hasn't", "has not"), ("hadn't", "had not"),
            
            ("we'd", "we would"), ("he'd", "he would"), ("you'd", "you would"), ("they'd", "they would"), 
            ("who'd", "who would"), ("wouldn't", "would not"),
            
            ("i'll", "i will"), ("you'll", "you will"), ("he'll", "he will"), ("she'll", "she will"),
            ("it'll", "it will"), ("we'll", "we will"), ("they'll", "they will"),
            
            ("i've", "i have"), ("you've", "you have"), ("we've", "we have"), ("they've", "they've"),
            ("&", " and ")
           ]

def sub_contractions(text):
    new_text = text
    for r in replaces:
        new_text = re.sub(f"{r[0]}", f"{r[1]}", new_text)
    return new_text

lemmatizer = WordNetLemmatizer()
def preprocess_text(text):
    if type(text) is float:  #Comprobando valores nulos, expresados como un float NaN
        ret_text = []
    else:
        new_text = text.lower()
        new_text = sub_contractions(new_text)
        new_text = re.sub("[^A-Za-zÀ-ÖØ-öø-ÿ0-9 ]", "", new_text)
        
        tok_text = word_tokenize(new_text)
        stop_words = list(stopwords.words("english"))
        tok_text = [word for word in tok_text if word not in stop_words]
        tok_text = [lemmatizer.lemmatize(word) for word in tok_text]
        ret_text = tok_text
        
        """ Processing feedback
        joint_text = " ".join(tok_text)
        print(new_text)
        print(joint_text)
        print() #"""
        
    return ret_text

'''TITLE'''

df["title"] = df["title"].apply(preprocess_text)
df["title"].head()

### 1.3.3 Tagline

Para la columna ``tagline``, la cual es de tipo texto, también se deberá llevar a cabo procesamiento de lenguaje natural.



'''TAGLINE'''

df["tagline"] = df["tagline"].apply(preprocess_text)
df["tagline"].head()

### 1.3.4 Overview

También con la columna ``overview``, se debe llevar a cabo de la misma forma.

'''OVERVIEW'''

df["overview"] = df["overview"].apply(preprocess_text)
df["overview"].head()

### 1.3.5 Vote count

Para liminamos las filas con valores nulos. Si el numero de votos es menor a 1 no es útil, puesto que una pelicula sin votos no sirve para deducir en función de la media de votos de una pelicula.

'''VOTE_COUNT'''

df["vote_count"] = df.drop(df[df.vote_count.isna()].index, inplace=False)["vote_count"].astype("int32")
df = df[df.vote_count >= 1]
df["vote_count"].head()

### 1.3.6 Vote average

Eliminamos las filas con valores nulos de la columna ``vote_average``. Las filas anteriores, es decir, que tengan votos significativos, considerar todos los votos (hay filas que tienen una media de 0 pero ningun voto, por lo que no son válidas)

'''VOTE_AVERAGE'''

df["vote_average"] = df.drop(df[df.vote_average.isna()].index, inplace=False)["vote_average"]
df["vote_average"].head()

### 1.3.7 Release date

Eliminamos las filas con valores nulos de esta columna. La columna de las fechas estaba en string (probablemente debido a los valores erroneos que había al principio) así que la convertimos en una columna de tipo datetime.

Podemos dejarlos como están o bien agrupar los años para extraer conclusiones

'''RELEASE_DATE'''

df["release_date"] = df.drop(df[df.release_date.isna()].index, inplace=False)["release_date"]
df["release_date"] = pd.to_datetime(df["release_date"])
df["release_date"].head()

#df_years = preprocessed['release_date'].groupby(preprocessed.release_date.dt.year)
#print(df_years.first())

df["release_month"] = df["release_date"].dt.month
df["release_day"] = df["release_date"].dt.day_name()
df.drop(["release_date"], axis=1, inplace=True)
df.head()

### 1.3.8 Runtime

Eliminamos las filas con valores nulos y se borran los que tienen valor 0 porque no tiene sentido que una pelicula tenga 0 minutos.

'''RUNTIME'''

df["runtime"] = df.drop(df[df.runtime.isna()].index, inplace=False)["runtime"]
df = df[df.runtime >= 1]
df["runtime"].head()

### 1.3.9 Original language

Para elegir los idiomas validos vamos a ver los 6 más usados y juntar los otros en la categoria resto.


Cambiar filas con idiomas inválidos.

'''ORIGINAL LANGUAGE'''

idiomas_validos = {'en', 'fr', 'it', 'ja', 'de', 'es'}

def preprocess_language(text):
    if text in idiomas_validos:
        return text
    else:
        return "other"

df["original_language"] = df["original_language"].apply(preprocess_language)
df["original_language"].value_counts()

Para elegir los idiomas validos vamos a ver los 6 más usados y hacer un one hot encoding con los que aparezcan.

def language_counts(rows):
    languages_dict = {}
    
    for row in rows:
        try:
            languages = ast.literal_eval(row)
            
            for lang in languages:
                code = lang["iso_639_1"]
                
                if code in languages_dict.keys():
                    languages_dict[code] += 1
                else:
                    languages_dict[code] = 1
        except Exception as e:
            continue
            
    sorted_counts = sorted(languages_dict.items(), key=lambda x: x[1], reverse=True)
    return sorted_counts



def treat_languages(df, best_languages):
    for lang in best_languages:
        df[f'spoken_lang_{lang}'] = 0
    df[f'spoken_lang_other'] = 0
    
    for i, row in df.iterrows():
        try:
            language_list = ast.literal_eval(row['spoken_languages'])
            for language_dict in language_list:
                code = language_dict['iso_639_1']
                if code in best_languages:
                    df.iloc[i, df.columns.get_loc(f'spoken_lang_{code}')] = 1
                else:
                    df.iloc[i, df.columns.get_loc(f'spoken_lang_other')] = 1
        except:
            continue


lang_counts = language_counts(df["spoken_languages"])
lang_valid = [lang_counts[i][0] for i in range(7)]
lang_valid.append("other")

treat_languages(df, lang_valid)
df.drop(['spoken_languages'], axis=1, inplace=True)
df.head()

### 1.3.10 Genre

Para elegir los generos validos vamos a ver los 5 mas usados y juntar los otros en la categoria resto.

"""

#print(pd.isnull(df).any())
print(df["genres"].value_counts(dropna=False))

#Camabiar filas con generos inválidos
generos_validos = {'1', '2', '3', '4', '5'}

for nodo in df:
   if genres not in generos_validos 
      genres = other

"""

def genre_counts(rows):
    genre_dict = {}
    
    for row in rows:
        try:
            genres = ast.literal_eval(row)
            
            for genre in genres:
                name = genre["name"]
                
                if name in genre_dict.keys():
                    genre_dict[name] += 1
                else:
                    genre_dict[name] = 1
        except Exception as e:
            continue
            
    sorted_counts = sorted(genre_dict.items(), key=lambda x: x[1], reverse=True)
    return sorted_counts



def treat_genres(df, valid_genres):
    for genre in valid_genres:
        df[f'genre_{genre}'] = 0
    
    df[f'genre_other'] = 0
    
    for i, row in df.iterrows():
        try:
            genre_list = ast.literal_eval(row['genres'])
            for genre_dict in genre_list:
                name = genre_dict['name']
                
                if name in valid_genres:
                    df.iloc[i, df.columns.get_loc(f'genre_{name}')] = 1
                else:
                    df.iloc[i, df.columns.get_loc(f'genre_other')] = 1
        except:
            continue

genres = genre_counts(df["genres"])
genre_valid = [genres[i][0] for i in range(7)]
genre_valid.append("other")

treat_genres(df, genre_valid)
df.drop(['genres'], axis=1, inplace=True)
df.head()

#Tenemos que elegir entre tocar o no el presupuesto, si lo dejamos tal cual o en 4 categorias que habia dudas 
#(que sean super produccion, gran produccion, produccion normal y produccion bajo presupuesto por ejemplo)

### 1.3.11 Production companies

En cuanto a la columna `production_companies`, esta es una columna con diccionarios de compañías, por lo que se hará será coger el 'id' de cada compañía productora de la película y contar las apariciones de cada compañía para, posteriormente, en esa misma columna poner el valor de las ocurrencias (popularidad) máxima de las compañías involucradas.

'''PRODUCTION_COMPANIES'''

def popularity_companies(rows):
    companies_popularity = {}
    for row in rows:
        try:
            companies = ast.literal_eval(row)
            for company in companies:
                if company['id'] in companies_popularity.keys():
                    companies_popularity[company['id']] += 1
                else:
                    companies_popularity[company['id']] = 1
        except Exception:
            continue
    return companies_popularity

companies_popularity = popularity_companies(df['production_companies'])

def treat_companies(row):
    max_popularity = 0
    try:
        companies = ast.literal_eval(row)
        for company in companies:
            if int(companies_popularity[company['id']]) > max_popularity:
                max_popularity = companies_popularity[company['id']]
        return max_popularity
    except Exception:
        return max_popularity

df['production_companies'] = df['production_companies'].apply(treat_companies)
df['production_companies'].head(20)

### 1.3.12 Production countries

Para `production_countries`, esta es una columna de diccionarios de compañías, por lo que el proceso sera coger los id's de cada compañía y crear una lista de los siete países que más aparecen. Esto se convertirá en columna de one hot encoding.

'''PRODUCTION_COUNTRIES'''

def popularity_countries(rows):
    countries_popularity = {}
    
    for row in rows:
        try:
            countries = ast.literal_eval(row)
            for country in countries:
                if country['iso_3166_1'] in countries_popularity.keys():
                    countries_popularity[country['iso_3166_1']] += 1
                else:
                    countries_popularity[country['iso_3166_1']] = 1
        except:
            continue
            
    countries = sorted(countries_popularity.items(), key=lambda x: x[1], reverse=True)[:7]
    best_countries = []
    
    for country in countries:
        best_countries.append(country[0])
        
    return best_countries


def treat_countries(df, best_countries):
    for country in best_countries:
        df[f'production_{country}'] = 0
    df['production_other'] = 0
        
    for i, row in df.iterrows():
        try:
            countries = ast.literal_eval(row['production_countries'])
            for country in countries:
                if country['iso_3166_1'] in best_countries:
                    df.iloc[i, df.columns.get_loc(f'production_{country["iso_3166_1"]}')] = 1
                else:
                    df.iloc[i, df.columns.get_loc('production_other')] = 1
        except:
            continue

popular_countries = popularity_countries(df['production_countries'])

treat_countries(df, popular_countries)
df.drop(['production_countries'], axis=1, inplace=True)
df.head()

### 1.3.13 Budget

df["budget"] = df.drop(df[df.budget.isna()].index, inplace=False)["budget"]
df["budget"].head()

### 1.3.14 Belongs to collection

def preprocess_collection(coleccion):
    id_coleccion = -1
    if type(coleccion) != float:
        id_coleccion = re.search("'id' *: *([0-9]+)", coleccion)[1]
    return id_coleccion

df["belongs_to_collection"] = df["belongs_to_collection"].apply(preprocess_collection)
df["belongs_to_collection"]

### 1.3.13 Revenue

Dado que para está característica se ha podido observar que faltan bastantes datos, esto nos servirá para crear dos datasets. Se tendrá un dataset con la columna de ``revenue`` con ningún valor nulo para el entrenamiento, y otro dataset sin esta columna que sería para el training.

'''REVENUE'''

revenue_valor = df[df["revenue"] > 0]
revenue_cero = df[df["revenue"] == 0]

print(len(revenue_valor))
print(len(revenue_cero))

## 1.5 Exportación de datos preprocesados

Una vez prerporcesadas todas las columnas, será turno de crear otros datasets que serán utilizados en nuestro sistema de predicción. Se tendrá unos datasets sin ratings y otros con ratings, pero con menos samples.

revenue_valor.to_csv("movies_dataset/preprocessed_training.csv", index = False)

revenue_cero.to_csv("movies_dataset/preprocessed_testing.csv", index = False)
