==> 2_Transformation.ipynb <==
# 2. Transformación

## 2.1 Importando paquetes y datasets

import pandas as pd
import nltk
import math
import io
import re
import ast
import warnings
warnings.filterwarnings('ignore')

#Imports de vectorizacion
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.preprocessing import LabelEncoder

# with open("res/preprocesado_training.csv", "r", encoding='utf-8') as f:
#     df_traing = pd.read_csv(f, encoding="utf-8")
#     print("Dataset loaded")

# with open("res/preprocesado_testing.csv", "r", encoding='utf-8') as f:
#     df_test = pd.read_csv(f, encoding="utf-8")
#     print("Dataset loaded")

with open("movies_dataset/preprocessed_training.csv", "r", encoding='utf-8') as f:
    df_train = pd.read_csv(f, encoding="utf-8")
    print("Training dataset loaded")

with open("movies_dataset/preprocessed_testing.csv", "r", encoding='utf-8') as f:
    df_test = pd.read_csv(f, encoding="utf-8")
    print("Testing dataset loaded")

## 2.2 Transformación de lenguaje natural

### 2.2.1 Transformación de los datos de training

Una vez realizado el tokenizado de las columnas compuestas por texto plano, la siguiente labor será la de vectorizar estas columnas, de forma que un algoritmo pueda procesar estos datos. Para ello se cogerán todas aquellas columnas de lenguaje natural y se tomarán todas las palabras aparecidas en ellas.

def to_plain(row):
    text = ''
    items = ast.literal_eval(row)
    items = set(items)
    for item in items:
        text += f' {item}'
    return text

title = df_train['title'].apply(to_plain)
title.head()

tagline = df_train['tagline'].apply(to_plain)
tagline.head()

overview = df_train['overview'].apply(to_plain)
overview.head()

Una vez pasado a texto plano todas las columnas, se vectorizarán cada una de estas características de los datos para training.

def best_words(raw):
    vec = CountVectorizer().fit(raw)
    words = vec.transform(raw)
    sum_words = words.sum(axis=0) 
    words_freq = [(word, sum_words[0, idx]) for word, idx in vec.vocabulary_.items()]
    words_freq =sorted(words_freq, key = lambda x: x[1], reverse=True)
    
    n = int(words_freq[0][1] * 0.3)
    best_words = []
    for word in words_freq:
        if word[1] >= n:
            best_words.append(word[0])
        else:
            break
    return best_words

overview_words = best_words(overview)
title_words= best_words(title)
tagline_words = best_words(tagline)

def treat_text(df, best_words, column):
    for word in best_words:
        df[f'{column}_{word}'] = 0
        
    for i, row in df.iterrows():
        try:
            items = ast.literal_eval(row[column])
            for word in items:
                if word in best_words:
                    df.iloc[i, df.columns.get_loc(f'{column}_{word}')] = 1
        except:
            continue

Las palabras con más ocurrencias de cada características serán añadidas como columnas.

treat_text(df_train, overview_words, 'overview')

treat_text(df_train, title_words, 'title')

treat_text(df_train, tagline_words, 'tagline')

df_train.head()

df_train.drop(['title', 'tagline', 'overview'], axis=1, inplace=True)

### 2.2.2 Transformación de los datos de testing

Para el dataset para testing, también se añadirán estas mismas columnas con One Hot Encoding, si aparacen en cada una de las columnas ya mencionadas anteriormente.

treat_text(df_test, overview_words, 'overview')

treat_text(df_test, title_words, 'title')

treat_text(df_test, tagline_words, 'tagline')

df_test.drop(['title', 'tagline', 'overview'], axis=1, inplace=True)

## 2.3 Transformación de columnas categóricas

### 2.3.1 Release day of the week

Esta columna que indica el día de la semana, será procesada mediante Label Encoding, tanto del training y del testing.

lb_enc = LabelEncoder()

df_train['release_day'] = lb_enc.fit_transform(df_train['release_day'])
df_test['release_day'] = lb_enc.fit_transform(df_test['release_day'])

### 2.3.2 Original language

Esta columna también será procesada mediante Label Encoding.

df_train['original_language'] = lb_enc.fit_transform(df_train['original_language'])
df_test['original_language'] = lb_enc.fit_transform(df_test['original_language'])

## 2.4 Actualización de los datasets

df_train.to_csv("movies_dataset/transformed_training.csv", index = False)

df_test.to_csv("movies_dataset/transformed_testing.csv", index = False)
