==> 3_Mineria.ipynb <==
from sklearn.model_selection import train_test_split
from sklearn.ensemble import RandomForestRegressor
from sklearn.metrics import mean_absolute_error
import pandas as pd
import numpy as np

with open("movies_dataset/transformed_training.csv", "r") as f:
    df = pd.read_csv(f)

##Prediccion de la media de votos

#Eliminacion de columnas no conocidas y separacion en entrenamiento y validacion
df_votes = df.drop(["vote_count", "revenue", "id"], axis=1)
train_votes, val_votes = train_test_split(df_votes, test_size=0.2, random_state=0)


#Obtencion de las columnas mas correlacionadas
mtx_corr = train_votes.corr()
minimum_correlation_votes = 0  #Todas las features
revenue_correlations_votes = mtx_corr["vote_average"][abs(mtx_corr["vote_average"]) > minimum_correlation_votes]

print(revenue_correlations_votes)

selected_columns_votes = list(revenue_correlations_votes.index)
selected_columns_votes.remove("vote_average")


#Seleccion de features y objetivos
features_train_votes = train_votes[selected_columns_votes]
goal_train_votes = train_votes[["vote_average"]]

features_val_votes = val_votes[selected_columns_votes]
goal_val_votes = val_votes[["vote_average"]]


#Seleccion del modelo y optimizacion de hiperparametros
model_votes = RandomForestRegressor(random_state=0)
model_votes.fit(features_train_votes, goal_train_votes)


#Prueba del modelo sobre ambos conjuntos para ver el error
predictions_train_votes = model_votes.predict(features_train_votes)
predictions_val_votes = model_votes.predict(features_val_votes)

MAE_train_votes = mean_absolute_error(goal_train_votes, predictions_train_votes)
MAE_val_votes = mean_absolute_error(goal_val_votes, predictions_val_votes)

print()
print("Training error:", MAE_train_votes)
print("Validation error:", MAE_val_votes)

##Prediccion de la recaudacion

#Eliminacion de columnas no conocidas y separacion en entrenamiento y validacion
df_revenue = df.drop(["vote_count", "vote_average", "id"], axis=1)
train_revenue, val_revenue = train_test_split(df_revenue, test_size=0.2, random_state=0)


#Obtencion de las columnas mas correlacionadas
mtx_corr = train_revenue.corr()
minimum_correlation_revenue = 0.01 #Con correlacion mayor que 0.01
revenue_correlations_revenue = mtx_corr["revenue"][abs(mtx_corr["revenue"]) > minimum_correlation_revenue]

print(revenue_correlations_revenue)

selected_columns_revenue = list(revenue_correlations_revenue.index)
selected_columns_revenue.remove("revenue")


#Seleccion de features y objetivos
features_train_revenue = train_revenue[selected_columns_revenue]
goal_train_revenue = train_revenue[["revenue"]]

features_val_revenue = val_revenue[selected_columns_revenue]
goal_val_revenue = val_revenue[["revenue"]]


#Seleccion del modelo y optimizacion de hiperparametros
model_revenue = RandomForestRegressor(random_state=0)
model_revenue.fit(features_train_revenue, goal_train_revenue)


#Prueba del modelo sobre ambos conjuntos para ver el error
predictions_train_revenue = model_revenue.predict(features_train_revenue)
predictions_val_revenue = model_revenue.predict(features_val_revenue)

MAE_train_revenue = mean_absolute_error(goal_train_revenue, predictions_train_revenue)
MAE_val_revenue = mean_absolute_error(goal_val_revenue, predictions_val_revenue)

print()
print("Training error:", MAE_train_revenue)
print("Validation error:", MAE_val_revenue)

# 0.05 40M
# 0.01 39.5M
# 0.1  39.6M

### SECCIÓN DE APLICACIÓN DEL MODELO

predecir_dict = {
    19995 : "Avatar",
    597 : "Titanic",
    109445 : "Frozen",
    862 : "Toy Story 1",
    863 : "Toy Story 2",
    27205 : "Inception"
}

predecir = df[df["id"].isin(predecir_dict)]

predecir_votes = predecir[selected_columns_votes]
predecir_revenue = predecir[selected_columns_revenue]

predictions_votes = model_votes.predict(predecir_votes)
predictions_revenue = model_revenue.predict(predecir_revenue)

results = pd.DataFrame({
    "Pelicula" : list(predecir_dict.values()),
    "Prediccion votacion" : predictions_votes,
    "Votacion real" :  predecir["vote_average"],
    "Prediccion recaudacion" : predictions_revenue,
    "Recaudacion real" : predecir["revenue"]
})

results
